+++ constants/nerdfonts/nerdfonts.go
@@ -3,9 +3,12 @@ package nerdfonts
 const (
 	Time           = ""
 	Author         = ""
-	Score          = "ﰵ"
+	Score          = "󰜷"
 	Comment        = ""
 	Tag            = ""
+	Document       = ""
+	Audio          = ""
+	Video          = ""
 	LeftSeparator  = ""
 	RightSeparator = ""
 )