+++ app/app.go
@@ -3,5 +3,5 @@ package app
 const (
 	Name               = "circumflex"
 	Version            = "2.9.1-dev"
-	MinimumLessVersion = 608
+	MinimumLessVersion = 633
 )