+++ app/app.go
@@ -2,6 +2,6 @@ package app
 
 const (
 	Name               = "circumflex"
-	Version            = "2.9.1-dev"
+	Version            = "2.9.1"
 	MinimumLessVersion = 633
 )