+++ CHANGELOG.md
@@ -6,6 +6,7 @@ _WIP_
 **Bugfixes**
 - Fixed a bug where `less` would open the comment section on the second line
 - Fixed a bug where `less` would not properly render Nerd Fonts
+- Updated Nerd Font icons to version 3
 
 ## 2.9
 _05.05.23_