+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 2.9.1
 _WIP_
 
+**Core**
+- Set minimum `less` version to latest stable release (633)
+
 **Bugfixes**
 - Fixed a bug where `less` would open the comment section on the second line
 - Fixed a bug where `less` would not properly render Nerd Fonts